requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.1
sqlalchemy>=2.0.27
pymysql>=1.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.1
pandas>=2.2.2
openpyxl>=3.1.2
python-multipart>=0.0.9
slowapi>=0.1.9
pydantic>=2.6.1
pydantic-settings>=2.1.0
httpx>=0.27.0
Pillow>=10.2.0
razorpay>=1.4.1
gunicorn>=21.2.0
